Refactor stylelint plugin to ESM
