+++ stylelint-plugins/machinery/css.js
@@ -1,12 +1,10 @@
-module.exports = {
-  flexChildProps: [
-    'flex', 'flex-grow', 'flex-shrink', 'flex-basis',
-  ],
-  gridChildProps: [
-    'grid', 'grid-area', 'grid-column', 'grid-row',
-    'grid-column-start', 'grid-column-end', 'grid-row-start', 'grid-row-end',
-  ],
-  flexOrGridChildProps: [
-    'order'
-  ],
-}
+export const flexChildProps = [
+  'flex', 'flex-grow', 'flex-shrink', 'flex-basis',
+]
+export const gridChildProps = [
+  'grid', 'grid-area', 'grid-column', 'grid-row',
+  'grid-column-start', 'grid-column-end', 'grid-row-start', 'grid-row-end',
+]
+export const flexOrGridChildProps = [
+  'order'
+]
